refactor(prompts): move DebatePromptBuilder to src/prompts/builders/ (phase 2)
Move src/debate/prompt-builder.ts → src/prompts/builders/debate-builder.ts.
All internal logic is preserved byte-identical; only import paths are updated.

Circular dependency avoided: debate-builder.ts imports directly from
src/debate/personas.ts and src/debate/types.ts (leaf modules), not from
the src/debate barrel which re-exports DebatePromptBuilder from src/prompts.

Callsites updated:
- session-hybrid, session-one-shot, session-plan, session-stateful: ../prompts
- src/review/dialogue.ts: ../prompts
- src/debate/index.ts: re-exports from ../prompts (backward-compat preserved)
- test/unit/debate/prompt-builder.test.ts: src/prompts barrel
