+++ src/debate/index.ts
@@ -6,8 +6,8 @@ export { DebateSession } from "./session";
 export { _debateSessionDeps, resolveDebaterModel } from "./session-helpers";
 export type { DebateSessionOptions, ResolverContextInput, ResolveOutcome } from "./session-helpers";
 export { majorityResolver, synthesisResolver, judgeResolver } from "./resolvers";
-export { DebatePromptBuilder } from "./prompt-builder";
-export type { StageContext, PromptBuilderOptions, ReviewStoryContext } from "./prompt-builder";
+export { DebatePromptBuilder } from "../prompts";
+export type { StageContext, PromptBuilderOptions, ReviewStoryContext } from "../prompts";
 export { PERSONA_FRAGMENTS, buildDebaterLabel, buildPersonaBlock, resolvePersonas } from "./personas";
 export type { DebaterPersona } from "./types";
 export type {
