+++ src/debate/session-one-shot.ts
@@ -5,9 +5,9 @@
  */
 
 import type { NaxConfig } from "../config";
+import { DebatePromptBuilder } from "../prompts";
 import { allSettledBounded } from "./concurrency";
 import { buildDebaterLabel, resolvePersonas } from "./personas";
-import { DebatePromptBuilder } from "./prompt-builder";
 import {
   type ResolveOutcome,
   type ResolvedDebater,
