+++ src/review/dialogue.ts
@@ -9,10 +9,10 @@ import type { SemanticVerdict } from "../acceptance/types";
 import type { AgentAdapter } from "../agents/types";
 import type { NaxConfig } from "../config";
 import { resolveModelForAgent } from "../config/schema-types";
-import { DebatePromptBuilder } from "../debate/prompt-builder";
 import type { DebateResolverContext } from "../debate/types";
 import { NaxError } from "../errors";
 import type { ReviewFinding } from "../plugins/types";
+import { DebatePromptBuilder } from "../prompts";
 import { parseLLMJson, tryParseLLMJson } from "../utils/llm-json";
 import type { SemanticStory } from "./semantic";
 import type { SemanticReviewConfig } from "./types";
